+++ CTFd/api/v1/files.py
@@ -7,11 +7,12 @@ from CTFd.api.v1.helpers.request import validate_args
 from CTFd.api.v1.helpers.schemas import sqlalchemy_to_pydantic
 from CTFd.api.v1.schemas import APIDetailedSuccessResponse, APIListSuccessResponse
 from CTFd.constants import RawEnum
-from CTFd.models import Files, db
+from CTFd.models import Challenges, Files, db
 from CTFd.schemas.files import FileSchema
 from CTFd.utils import uploads
-from CTFd.utils.decorators import admins_only
+from CTFd.utils.decorators import access_granted_only
 from CTFd.utils.helpers.models import build_model_filters
+from CTFd.utils.user import has_right_or_is_author
 
 files_namespace = Namespace("files", description="Endpoint to retrieve Files")
 
@@ -37,13 +38,13 @@ files_namespace.schema_model(
 
 @files_namespace.route("")
 class FilesList(Resource):
-    @admins_only
+    @access_granted_only("api_files_list_get")
     @files_namespace.doc(
         description="Endpoint to get file objects in bulk",
         responses={
             200: ("Success", "FileListSuccessResponse"),
             400: (
-                "An error occured processing the provided or stored data",
+                "An error occurred processing the provided or stored data",
                 "APISimpleErrorResponse",
             ),
         },
@@ -74,13 +75,13 @@ class FilesList(Resource):
 
         return {"success": True, "data": response.data}
 
-    @admins_only
+    @access_granted_only("api_files_list_post")
     @files_namespace.doc(
         description="Endpoint to get file objects in bulk",
         responses={
             200: ("Success", "FileDetailedSuccessResponse"),
             400: (
-                "An error occured processing the provided or stored data",
+                "An error occurred processing the provided or stored data",
                 "APISimpleErrorResponse",
             ),
         },
@@ -107,38 +108,43 @@ class FilesList(Resource):
 
 @files_namespace.route("/<file_id>")
 class FilesDetail(Resource):
-    @admins_only
+    @access_granted_only("api_files_detail_get")
     @files_namespace.doc(
         description="Endpoint to get a specific file object",
         responses={
             200: ("Success", "FileDetailedSuccessResponse"),
             400: (
-                "An error occured processing the provided or stored data",
+                "An error occurred processing the provided or stored data",
                 "APISimpleErrorResponse",
             ),
         },
     )
     def get(self, file_id):
         f = Files.query.filter_by(id=file_id).first_or_404()
-        schema = FileSchema()
-        response = schema.dump(f)
+        challenge = Challenges.query.filter_by(id=f.challenge_id).first_or_404()
+        if has_right_or_is_author("api_files_detail_get", challenge.author_id):
+            schema = FileSchema()
+            response = schema.dump(f)
 
-        if response.errors:
-            return {"success": False, "errors": response.errors}, 400
+            if response.errors:
+                return {"success": False, "errors": response.errors}, 400
 
-        return {"success": True, "data": response.data}
+            return {"success": True, "data": response.data}
+        return {"success": False}
 
-    @admins_only
+    @access_granted_only("api_files_detail_delete")
     @files_namespace.doc(
         description="Endpoint to delete a file object",
         responses={200: ("Success", "APISimpleSuccessResponse")},
     )
     def delete(self, file_id):
         f = Files.query.filter_by(id=file_id).first_or_404()
+        challenge = Challenges.query.filter_by(id=f.challenge_id).first_or_404()
+        if has_right_or_is_author("api_files_detail_delete", challenge.author_id):
+            uploads.delete_file(file_id=f.id)
+            db.session.delete(f)
+            db.session.commit()
+            db.session.close()
 
-        uploads.delete_file(file_id=f.id)
-        db.session.delete(f)
-        db.session.commit()
-        db.session.close()
-
-        return {"success": True}
+            return {"success": True}
+        return {"success": False}
